feat: include `createLogger` function in core export
