+++ .gitignore
@@ -1,14 +1,9 @@
 /dist
-!/dist/logger.d.ts
 /docs/.vuepress/dist
-/examples/**/build.js
 /coverage
 /docs/.vuepress/dist
-/examples/**/build.js
 /test/e2e/reports
 /test/e2e/screenshots
-/types/typings
-/types/test/*.js
 *.log
 .DS_Store
 node_modules
