+++ examples/classic/chat/store/index.js
@@ -1,8 +1,7 @@
-import { createStore } from 'vuex'
+import { createStore, createLogger } from 'vuex'
 import * as getters from './getters'
 import * as actions from './actions'
 import mutations from './mutations'
-import createLogger from '../../../../src/plugins/logger'
 
 const state = {
   currentThreadID: null,
