+++ examples/classic/shopping-cart/store/index.js
@@ -1,7 +1,6 @@
-import { createStore } from 'vuex'
+import { createStore, createLogger } from 'vuex'
 import cart from './modules/cart'
 import products from './modules/products'
-import createLogger from '../../../../src/plugins/logger'
 
 const debug = process.env.NODE_ENV !== 'production'
