+++ examples/composition/todomvc/store/plugins.js
@@ -1,5 +1,5 @@
+import { createLogger } from 'vuex'
 import { STORAGE_KEY } from './mutations'
-import createLogger from '../../../../src/plugins/logger'
 
 const localStoragePlugin = store => {
   store.subscribe((mutation, { todos }) => {
