+++ package.json
@@ -17,9 +17,7 @@
   ],
   "scripts": {
     "dev": "node examples/server.js",
-    "build": "npm run build:main && npm run build:logger",
-    "build:main": "node scripts/build-main.js",
-    "build:logger": "node scripts/build-logger.js",
+    "build": "node scripts/build.js",
     "lint": "eslint src test",
     "test": "npm run lint && npm run test:types && npm run test:unit && npm run test:ssr && npm run test:e2e",
     "test:unit": "jest --testPathIgnorePatterns test/e2e",
