--- scripts/build-logger.js
@@ -1,5 +1,0 @@
-const { run } = require('./build')
-
-const files = ['dist/logger.js']
-
-run('rollup.logger.config.js', files)
