+++ src/index.cjs.js
@@ -1,7 +1,7 @@
 import { createStore, Store } from './store'
 import { useStore } from './injectKey'
 import { mapState, mapMutations, mapGetters, mapActions, createNamespacedHelpers } from './helpers'
-import createLogger from './plugins/logger'
+import { createLogger } from './plugins/logger'
 
 export default {
   version: '__VERSION__',
